clean up acceptance suites
