--- acceptance/openstack/identity/v2/pkg.go
@@ -1,1 +1,0 @@
-package v2
