+++ acceptance/openstack/networking/v1/vpc_test.go
@@ -3,6 +3,7 @@ package v1
 import (
 	"testing"
 
+	"github.com/huaweicloud/golangsdk"
 	"github.com/huaweicloud/golangsdk/acceptance/clients"
 	"github.com/huaweicloud/golangsdk/acceptance/tools"
 	"github.com/huaweicloud/golangsdk/openstack/networking/v1/vpcs"
@@ -31,11 +32,11 @@ func TestVpcsCRUD(t *testing.T) {
 	}
 
 	// Create a vpc
-	vpc, err := CreateVpc(t, client)
+	vpc, err := createVpc(t, client)
 	if err != nil {
 		t.Fatalf("Unable to create create: %v", err)
 	}
-	defer DeleteVpc(t, client, vpc.ID)
+	defer deleteVpc(t, client, vpc.ID)
 
 	tools.PrintResource(t, vpc)
 
@@ -55,4 +56,35 @@ func TestVpcsCRUD(t *testing.T) {
 	}
 
 	tools.PrintResource(t, newVpc)
+}
+
+func createVpc(t *testing.T, client *golangsdk.ServiceClient) (*vpcs.Vpc, error) {
+
+	vpcName := tools.RandomString("TESTACC-", 8)
+
+	createOpts := vpcs.CreateOpts{
+		Name: vpcName,
+		CIDR: "192.168.20.0/24",
+	}
+
+	t.Logf("Attempting to create vpc: %s", vpcName)
+
+	vpc, err := vpcs.Create(client, createOpts).Extract()
+	if err != nil {
+		return vpc, err
+	}
+	t.Logf("Created vpc: %s", vpcName)
+
+	return vpc, nil
+}
+
+func deleteVpc(t *testing.T, client *golangsdk.ServiceClient, vpcID string) {
+	t.Logf("Attempting to delete vpc: %s", vpcID)
+
+	err := vpcs.Delete(client, vpcID).ExtractErr()
+	if err != nil {
+		t.Fatalf("Error deleting vpc: %v", err)
+	}
+
+	t.Logf("Deleted vpc: %s", vpcID)
 }
